Remove loadtest component
