+++ .gitfiles
@@ -100,7 +100,6 @@
 .github/workflows/dockers-index-deletion-image.yaml
 .github/workflows/dockers-index-operator-image.yaml
 .github/workflows/dockers-index-save-image.yaml
-.github/workflows/dockers-loadtest-image.yaml
 .github/workflows/dockers-manager-index-image.yaml
 .github/workflows/dockers-readreplica-rotate-image.yaml
 .github/workflows/dockers-release-branch-image.yaml
@@ -531,9 +530,6 @@ cmd/tools/benchmark/operator/main_test.go
 cmd/tools/benchmark/operator/sample.yaml
 cmd/tools/cli/benchmark/core/main.go
 cmd/tools/cli/benchmark/core/main_test.go
-cmd/tools/cli/loadtest/main.go
-cmd/tools/cli/loadtest/main_test.go
-cmd/tools/cli/loadtest/sample.yaml
 cmd/tools/cli/vdctl/main.go
 cmd/tools/cli/vdctl/main_test.go
 codecov.yaml
@@ -577,8 +573,6 @@ dockers/operator/helm/README.md
 dockers/template.md
 dockers/tools/benchmark/job/Dockerfile
 dockers/tools/benchmark/operator/Dockerfile
-dockers/tools/cli/loadtest/Dockerfile
-dockers/tools/cli/loadtest/README.md
 docs/api/README.md
 docs/api/build_proto.md
 docs/api/filter-gateway.md
@@ -608,7 +602,6 @@ docs/overview/component/mirror-gateway.md
 docs/overview/data-flow.md
 docs/performance/benchmark.md
 docs/performance/continuous-benchmark.md
-docs/performance/loadtest.md
 docs/performance/tuning-search-performance.md
 docs/support/FAQ.md
 docs/support/contacts.md
@@ -695,8 +688,16 @@ hack/benchmark/core/benchmark/strategy/util.go
 hack/benchmark/core/benchmark/strategy/util_test.go
 hack/benchmark/core/ngt/ngt_bench_test.go
 hack/benchmark/e2e/agent/core/ngt/ngt_bench_test.go
+hack/benchmark/internal/assets/assets.go
+hack/benchmark/internal/assets/assets_test.go
 hack/benchmark/internal/assets/dataset.go
 hack/benchmark/internal/assets/dataset_test.go
+hack/benchmark/internal/assets/hdf5_loader.go
+hack/benchmark/internal/assets/hdf5_loader_test.go
+hack/benchmark/internal/assets/large_dataset.go
+hack/benchmark/internal/assets/large_dataset_test.go
+hack/benchmark/internal/assets/small_dataset.go
+hack/benchmark/internal/assets/small_dataset_test.go
 hack/benchmark/internal/core/algorithm/algorithm.go
 hack/benchmark/internal/core/algorithm/ngt/ngt.go
 hack/benchmark/internal/core/algorithm/ngt/ngt_test.go
@@ -1566,9 +1567,6 @@ k8s/tools/benchmark/operator/crds/valdbenchmarkscenario.yaml
 k8s/tools/benchmark/operator/deployment.yaml
 k8s/tools/benchmark/operator/service.yaml
 k8s/tools/benchmark/operator/serviceaccount.yaml
-k8s/tools/cli/loadtest/configmap.yaml
-k8s/tools/cli/loadtest/cronjob.yaml
-k8s/tools/cli/loadtest/job.yaml
 pkg/agent/core/faiss/config/config.go
 pkg/agent/core/faiss/config/config_test.go
 pkg/agent/core/faiss/handler/doc.go
@@ -1935,26 +1933,6 @@ pkg/tools/benchmark/operator/service/option.go
 pkg/tools/benchmark/operator/service/option_test.go
 pkg/tools/benchmark/operator/usecase/benchmarkd.go
 pkg/tools/benchmark/operator/usecase/benchmarkd_test.go
-pkg/tools/cli/loadtest/assets/dataset.go
-pkg/tools/cli/loadtest/assets/dataset_test.go
-pkg/tools/cli/loadtest/assets/hdf5_loader.go
-pkg/tools/cli/loadtest/assets/hdf5_loader_test.go
-pkg/tools/cli/loadtest/assets/large_dataset.go
-pkg/tools/cli/loadtest/assets/large_dataset_test.go
-pkg/tools/cli/loadtest/assets/small_dataset.go
-pkg/tools/cli/loadtest/assets/small_dataset_test.go
-pkg/tools/cli/loadtest/config/config.go
-pkg/tools/cli/loadtest/config/config_test.go
-pkg/tools/cli/loadtest/service/insert.go
-pkg/tools/cli/loadtest/service/insert_test.go
-pkg/tools/cli/loadtest/service/loader.go
-pkg/tools/cli/loadtest/service/loader_option.go
-pkg/tools/cli/loadtest/service/loader_option_test.go
-pkg/tools/cli/loadtest/service/loader_test.go
-pkg/tools/cli/loadtest/service/search.go
-pkg/tools/cli/loadtest/service/search_test.go
-pkg/tools/cli/loadtest/usecase/load.go
-pkg/tools/cli/loadtest/usecase/load_test.go
 renovate.json
 rust/Cargo.lock
 rust/Cargo.toml
