+++ .github/labeler.yaml
@@ -153,15 +153,6 @@ area/filter/ingress:
           - apis/grpc/v1/payload/**/*
           - apis/proto/v1/filter/ingress/*
           - apis/proto/v1/payload/**/*
-area/tools/cli/loadtest:
-  - changed-files:
-      - any-glob-to-any-file:
-          - apis/proto/v1/agent/core/**/*
-          - apis/proto/v1/vald/**/*
-          - apis/grpc/v1/agent/core/**/*
-          - apis/grpc/v1/vald/**/*
-          - cmd/tools/cli/loadtest/**/*
-          - pkg/tools/cli/loadtest/**/*
 area/internal:
   - changed-files:
       - any-glob-to-any-file:
