+++ .github/workflows/dockers-image-scan.yaml
@@ -78,10 +78,6 @@ jobs:
     uses: ./.github/workflows/_docker-image-scan.yaml
     with:
       target: index-save
-  loadtest:
-    uses: ./.github/workflows/_docker-image-scan.yaml
-    with:
-      target: loadtest
   manager-index:
     uses: ./.github/workflows/_docker-image-scan.yaml
     with:
