--- .github/workflows/dockers-loadtest-image.yaml
@@ -1,222 +1,0 @@
-#
-# Copyright (C) 2019-2025 vdaas.org vald team <vald@vdaas.org>
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# You may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-# DO_NOT_EDIT this workflow file is generated by https://github.com/vdaas/vald/blob/main/hack/docker/gen/main.go
-
-name: "Build docker image: loadtest"
-on:
-  push:
-    branches:
-      - main
-      - release/v*.*
-      - "!release/v*.*.*"
-    tags:
-      - "*.*.*"
-      - "*.*.*-*"
-      - v*.*.*
-      - v*.*.*-*
-  pull_request:
-    paths:
-      - "!**/*_mock.go"
-      - "!**/*_test.go"
-      - .github/actions/docker-build/action.yaml
-      - .github/workflows/_docker-image.yaml
-      - .github/workflows/dockers-loadtest-image.yaml
-      - Makefile
-      - Makefile.d/**
-      - apis/grpc/v1/payload/*.go
-      - apis/grpc/v1/rpc/errdetails/*.go
-      - apis/grpc/v1/vald/*.go
-      - apis/proto/**
-      - cmd/tools/cli/loadtest/*.go
-      - dockers/tools/cli/loadtest/Dockerfile
-      - go.mod
-      - go.sum
-      - hack/benchmark/assets/x1b/*.go
-      - hack/docker/gen/main.go
-      - internal/backoff/*.go
-      - internal/cache/*.go
-      - internal/cache/cacher/*.go
-      - internal/cache/gache/*.go
-      - internal/circuitbreaker/*.go
-      - internal/config/*.go
-      - internal/conv/*.go
-      - internal/db/kvs/redis/*.go
-      - internal/db/nosql/cassandra/*.go
-      - internal/db/rdb/mysql/*.go
-      - internal/db/rdb/mysql/dbr/*.go
-      - internal/encoding/json/*.go
-      - internal/errors/*.go
-      - internal/file/*.go
-      - internal/info/*.go
-      - internal/io/*.go
-      - internal/k8s/*.go
-      - internal/log/*.go
-      - internal/log/format/*.go
-      - internal/log/glg/*.go
-      - internal/log/level/*.go
-      - internal/log/logger/*.go
-      - internal/log/nop/*.go
-      - internal/log/retry/*.go
-      - internal/log/zap/*.go
-      - internal/net/*.go
-      - internal/net/control/*.go
-      - internal/net/grpc/*.go
-      - internal/net/grpc/admin/*.go
-      - internal/net/grpc/codes/*.go
-      - internal/net/grpc/credentials/*.go
-      - internal/net/grpc/errdetails/*.go
-      - internal/net/grpc/health/*.go
-      - internal/net/grpc/interceptor/client/metric/*.go
-      - internal/net/grpc/interceptor/client/trace/*.go
-      - internal/net/grpc/interceptor/server/logging/*.go
-      - internal/net/grpc/interceptor/server/metric/*.go
-      - internal/net/grpc/interceptor/server/recover/*.go
-      - internal/net/grpc/interceptor/server/trace/*.go
-      - internal/net/grpc/keepalive/*.go
-      - internal/net/grpc/logger/*.go
-      - internal/net/grpc/pool/*.go
-      - internal/net/grpc/proto/*.go
-      - internal/net/grpc/reflection/*.go
-      - internal/net/grpc/status/*.go
-      - internal/net/grpc/types/*.go
-      - internal/net/http/rest/*.go
-      - internal/net/quic/*.go
-      - internal/observability/attribute/*.go
-      - internal/observability/metrics/*.go
-      - internal/observability/trace/*.go
-      - internal/os/*.go
-      - internal/params/*.go
-      - internal/rand/*.go
-      - internal/runner/*.go
-      - internal/safety/*.go
-      - internal/servers/server/*.go
-      - internal/strings/*.go
-      - internal/sync/*.go
-      - internal/sync/errgroup/*.go
-      - internal/sync/semaphore/*.go
-      - internal/sync/singleflight/*.go
-      - internal/timeutil/*.go
-      - internal/timeutil/location/*.go
-      - internal/tls/*.go
-      - internal/version/*.go
-      - pkg/tools/cli/loadtest/assets/*.go
-      - pkg/tools/cli/loadtest/config/*.go
-      - pkg/tools/cli/loadtest/service/*.go
-      - pkg/tools/cli/loadtest/usecase/*.go
-      - versions/GO_VERSION
-  pull_request_target:
-    types:
-      - opened
-      - reopened
-      - synchronize
-      - labeled
-    paths:
-      - "!**/*_mock.go"
-      - "!**/*_test.go"
-      - .github/actions/docker-build/action.yaml
-      - .github/workflows/_docker-image.yaml
-      - .github/workflows/dockers-loadtest-image.yaml
-      - Makefile
-      - Makefile.d/**
-      - apis/grpc/v1/payload/*.go
-      - apis/grpc/v1/rpc/errdetails/*.go
-      - apis/grpc/v1/vald/*.go
-      - apis/proto/**
-      - cmd/tools/cli/loadtest/*.go
-      - dockers/tools/cli/loadtest/Dockerfile
-      - go.mod
-      - go.sum
-      - hack/benchmark/assets/x1b/*.go
-      - hack/docker/gen/main.go
-      - internal/backoff/*.go
-      - internal/cache/*.go
-      - internal/cache/cacher/*.go
-      - internal/cache/gache/*.go
-      - internal/circuitbreaker/*.go
-      - internal/config/*.go
-      - internal/conv/*.go
-      - internal/db/kvs/redis/*.go
-      - internal/db/nosql/cassandra/*.go
-      - internal/db/rdb/mysql/*.go
-      - internal/db/rdb/mysql/dbr/*.go
-      - internal/encoding/json/*.go
-      - internal/errors/*.go
-      - internal/file/*.go
-      - internal/info/*.go
-      - internal/io/*.go
-      - internal/k8s/*.go
-      - internal/log/*.go
-      - internal/log/format/*.go
-      - internal/log/glg/*.go
-      - internal/log/level/*.go
-      - internal/log/logger/*.go
-      - internal/log/nop/*.go
-      - internal/log/retry/*.go
-      - internal/log/zap/*.go
-      - internal/net/*.go
-      - internal/net/control/*.go
-      - internal/net/grpc/*.go
-      - internal/net/grpc/admin/*.go
-      - internal/net/grpc/codes/*.go
-      - internal/net/grpc/credentials/*.go
-      - internal/net/grpc/errdetails/*.go
-      - internal/net/grpc/health/*.go
-      - internal/net/grpc/interceptor/client/metric/*.go
-      - internal/net/grpc/interceptor/client/trace/*.go
-      - internal/net/grpc/interceptor/server/logging/*.go
-      - internal/net/grpc/interceptor/server/metric/*.go
-      - internal/net/grpc/interceptor/server/recover/*.go
-      - internal/net/grpc/interceptor/server/trace/*.go
-      - internal/net/grpc/keepalive/*.go
-      - internal/net/grpc/logger/*.go
-      - internal/net/grpc/pool/*.go
-      - internal/net/grpc/proto/*.go
-      - internal/net/grpc/reflection/*.go
-      - internal/net/grpc/status/*.go
-      - internal/net/grpc/types/*.go
-      - internal/net/http/rest/*.go
-      - internal/net/quic/*.go
-      - internal/observability/attribute/*.go
-      - internal/observability/metrics/*.go
-      - internal/observability/trace/*.go
-      - internal/os/*.go
-      - internal/params/*.go
-      - internal/rand/*.go
-      - internal/runner/*.go
-      - internal/safety/*.go
-      - internal/servers/server/*.go
-      - internal/strings/*.go
-      - internal/sync/*.go
-      - internal/sync/errgroup/*.go
-      - internal/sync/semaphore/*.go
-      - internal/sync/singleflight/*.go
-      - internal/timeutil/*.go
-      - internal/timeutil/location/*.go
-      - internal/tls/*.go
-      - internal/version/*.go
-      - pkg/tools/cli/loadtest/assets/*.go
-      - pkg/tools/cli/loadtest/config/*.go
-      - pkg/tools/cli/loadtest/service/*.go
-      - pkg/tools/cli/loadtest/usecase/*.go
-      - versions/GO_VERSION
-jobs:
-  build:
-    uses: ./.github/workflows/_docker-image.yaml
-    with:
-      target: loadtest
-      platforms: linux/amd64
-    secrets: inherit
