+++ .github/workflows/update-deps.yaml
@@ -48,6 +48,7 @@ jobs:
           git_commit_gpgsign: true
       - name: Run update command
         run: |
+          make deps/install
           make files
           make update
           make helm/schema/all
@@ -57,9 +58,7 @@ jobs:
           make helm/docs/vald
           make helm/docs/vald-helm-operator
           make helm/docs/vald-benchmark-operator
-          : # delete `make license` if it is not needed.
-          make license
-          make format
+          make proto/replace license format format/go format/go/test
           make perm
       - name: Check difference
         id: check_diff
