--- cmd/tools/cli/loadtest/sample.yaml
@@ -1,78 +1,0 @@
-#
-# Copyright (C) 2019-2025 vdaas.org vald team <vald@vdaas.org>
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# You may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-version: v0.0.0
-time_zone: JST
-logging:
-  logger: glg
-  level: info
-  format: raw
-operation: insert
-dataset: fashion-mnist
-concurrency: 100
-batch_size: 10
-progress_duration: 3s
-addr: "localhost:8082"
-client:
-  addrs: []
-  health_check_duration: 1s
-  backoff:
-    backoff_factor: 0
-    backoff_time_limit: ""
-    enable_error_log: false
-    initial_duration: ""
-    jitter_limit: ""
-    maximum_duration: ""
-    retry_count: 0
-  call_option:
-    max_recv_msg_size: 0
-    max_retry_rpc_buffer_size: 0
-    max_send_msg_size: 0
-    wait_for_ready: true
-  connection_pool:
-    enable_dns_resolver: true
-  dial_option:
-    enable_backoff: false
-    initial_connection_window_size: 0
-    initial_window_size: 0
-    insecure: true
-    keepalive:
-      permit_without_stream: false
-      time: ""
-      timeout: ""
-    backoff_max_delay: ""
-    max_msg_size: 0
-    read_buffer_size: 0
-    tcp:
-      dialer:
-        dual_stack_enabled: true
-        keepalive: ""
-        timeout: ""
-      dns:
-        cache_enabled: false
-        cache_expiration: ""
-        refresh_duration: ""
-      tls:
-        ca: /path/to/ca
-        cert: /path/to/cert
-        enabled: false
-        key: /path/to/key
-    timeout: ""
-    write_buffer_size: 0
-  tls:
-    ca: /path/to/ca
-    cert: /path/to/cert
-    enabled: false
-    key: /path/to/key
