+++ hack/benchmark/internal/assets/assets_test.go
@@ -11,20 +11,29 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package main
+package assets
 
 // NOT IMPLEMENTED BELOW
 //
-// func Test_main(t *testing.T) {
-// 	type want struct{}
+// func TestData(t *testing.T) {
+// 	type args struct {
+// 		name string
+// 	}
+// 	type want struct {
+// 		want func(testing.TB) Dataset
+// 	}
 // 	type test struct {
 // 		name       string
+// 		args       args
 // 		want       want
-// 		checkFunc  func(want) error
-// 		beforeFunc func(*testing.T)
-// 		afterFunc  func(*testing.T)
+// 		checkFunc  func(want, func(testing.TB) Dataset) error
+// 		beforeFunc func(*testing.T, args)
+// 		afterFunc  func(*testing.T, args)
 // 	}
-// 	defaultCheckFunc := func(w want) error {
+// 	defaultCheckFunc := func(w want, got func(testing.TB) Dataset) error {
+// 		if !reflect.DeepEqual(got, w.want) {
+// 			return errors.Errorf("got: \"%#v\",\n\t\t\t\twant: \"%#v\"", got, w.want)
+// 		}
 // 		return nil
 // 	}
 // 	tests := []test{
@@ -32,12 +41,15 @@ package main
 // 		/*
 // 		   {
 // 		       name: "test_case_1",
+// 		       args: args {
+// 		           name:"",
+// 		       },
 // 		       want: want{},
 // 		       checkFunc: defaultCheckFunc,
-// 		       beforeFunc: func(t *testing.T,) {
+// 		       beforeFunc: func(t *testing.T, args args) {
 // 		           t.Helper()
 // 		       },
-// 		       afterFunc: func(t *testing.T,) {
+// 		       afterFunc: func(t *testing.T, args args) {
 // 		           t.Helper()
 // 		       },
 // 		   },
@@ -48,12 +60,15 @@ package main
 // 		   func() test {
 // 		       return test {
 // 		           name: "test_case_2",
+// 		           args: args {
+// 		           name:"",
+// 		           },
 // 		           want: want{},
 // 		           checkFunc: defaultCheckFunc,
-// 		           beforeFunc: func(t *testing.T,) {
+// 		           beforeFunc: func(t *testing.T, args args) {
 // 		               t.Helper()
 // 		           },
-// 		           afterFunc: func(t *testing.T,) {
+// 		           afterFunc: func(t *testing.T, args args) {
 // 		               t.Helper()
 // 		           },
 // 		       }
@@ -67,18 +82,18 @@ package main
 // 			tt.Parallel()
 // 			defer goleak.VerifyNone(tt, goleak.IgnoreCurrent())
 // 			if test.beforeFunc != nil {
-// 				test.beforeFunc(tt)
+// 				test.beforeFunc(tt, test.args)
 // 			}
 // 			if test.afterFunc != nil {
-// 				defer test.afterFunc(tt)
+// 				defer test.afterFunc(tt, test.args)
 // 			}
 // 			checkFunc := test.checkFunc
 // 			if test.checkFunc == nil {
 // 				checkFunc = defaultCheckFunc
 // 			}
 //
-// 			main()
-// 			if err := checkFunc(test.want); err != nil {
+// 			got := Data(test.args.name)
+// 			if err := checkFunc(test.want, got); err != nil {
 // 				tt.Errorf("error = %v", err)
 // 			}
 // 		})
