+++ hack/benchmark/internal/assets/hdf5_loader.go
