+++ hack/benchmark/internal/assets/small_dataset.go
